+++ src/main/java/com/example/demo/holding/HoldingService.java
@@ -3,7 +3,6 @@ package com.example.demo.holding;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -12,7 +11,7 @@ public class HoldingService {
     @Autowired
     HoldingRepository holdingRepository;
 
-    public void saveHolding(Long user_id, Long crypto_id, int quantity) {
+    public void saveHolding(Long user_id, Long crypto_id, double quantity) {
         HoldingDTO holding = new HoldingDTO(user_id, crypto_id, quantity);
         holdingRepository.save(holding, quantity);
     }
