+++ src/main/java/com/example/demo/transaction/Transaction.java
@@ -1,6 +1,5 @@
 package com.example.demo.transaction;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class Transaction {
@@ -8,14 +7,14 @@ public class Transaction {
     private Long user_id;
     private Long crypto_id;
     private String type;
-    private BigDecimal quantity;
-    private BigDecimal unit_price;
-    private BigDecimal total_cost;
+    private double quantity;
+    private double unit_price;
+    private double total_cost;
     private LocalDateTime created_at;
 
 
     public Transaction() {}
-    public Transaction(Long user_id, Long crypto_id, String type, BigDecimal quantity, BigDecimal unit_price) {
+    public Transaction(Long user_id, Long crypto_id, String type, double quantity, double unit_price) {
         this.user_id = user_id;
         this.crypto_id = crypto_id;
         this.type = type;
@@ -35,14 +34,14 @@ public class Transaction {
     public String getType() { return type; }
     public void setType(String type) { this.type = type; }
 
-    public BigDecimal getQuantity() { return quantity; }
-    public void setQuantity(BigDecimal quantity) { this.quantity = quantity; }
+    public double getQuantity() { return quantity; }
+    public void setQuantity(double quantity) { this.quantity = quantity; }
 
-    public BigDecimal getUnit_price() { return unit_price; }
-    public void setUnit_price(BigDecimal unit_price) { this.unit_price = unit_price; }
+    public double getUnit_price() { return unit_price; }
+    public void setUnit_price(double unit_price) { this.unit_price = unit_price; }
 
-    public BigDecimal getTotal_cost() { return total_cost; }
-    public void setTotal_cost(BigDecimal total_cost) { this.total_cost = total_cost; }
+    public double getTotal_cost() { return total_cost; }
+    public void setTotal_cost(double total_cost) { this.total_cost = total_cost; }
 
     public LocalDateTime getCreated_at() { return created_at; }
     public void setCreated_at(LocalDateTime created_at) { this.created_at = created_at; }
