+++ src/main/java/com/example/demo/transaction/TransactionDTO.java
@@ -1,16 +1,14 @@
 package com.example.demo.transaction;
 
-import java.math.BigDecimal;
-
 public class TransactionDTO {
     private Long user_id;
     private Long crypto_id;
     private String type;
-    private BigDecimal quantity;
-    private BigDecimal unit_price;
-    private BigDecimal total_cost;
+    private double quantity;
+    private double unit_price;
+    private double total_cost;
     public TransactionDTO() {}
-    public TransactionDTO(Long user_id, Long crypto_id, String type, BigDecimal quantity, BigDecimal unit_price, BigDecimal total_cost) {
+    public TransactionDTO(Long user_id, Long crypto_id, String type, double quantity, double unit_price, double total_cost) {
         this.user_id = user_id;
         this.crypto_id = crypto_id;
         this.type = type;
@@ -43,27 +41,27 @@ public class TransactionDTO {
         this.type = type;
     }
 
-    public BigDecimal getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(BigDecimal quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
 
-    public BigDecimal getUnit_price() {
+    public double getUnit_price() {
         return unit_price;
     }
 
-    public void setUnit_price(BigDecimal unit_price) {
+    public void setUnit_price(double unit_price) {
         this.unit_price = unit_price;
     }
 
-    public BigDecimal getTotal_cost() {
+    public double getTotal_cost() {
         return total_cost;
     }
 
-    public void setTotal_cost(BigDecimal total_cost) {
+    public void setTotal_cost(double total_cost) {
         this.total_cost = total_cost;
     }
 }
