+++ src/main/java/com/example/demo/transaction/TransactionRowMapper.java
@@ -13,9 +13,9 @@ public class TransactionRowMapper implements RowMapper<Transaction> {
         transaction.setUser_id(rs.getLong("user_id"));
         transaction.setCrypto_id(rs.getLong("crypto_id"));
         transaction.setType(rs.getString("type"));
-        transaction.setQuantity(rs.getBigDecimal("quantity"));
-        transaction.setUnit_price(rs.getBigDecimal("unit_price"));
-        transaction.setTotal_cost(rs.getBigDecimal("total_cost"));
+        transaction.setQuantity(rs.getDouble("quantity"));
+        transaction.setUnit_price(rs.getDouble("unit_price"));
+        transaction.setTotal_cost(rs.getDouble("total_cost"));
         transaction.setCreated_at(rs.getTimestamp("created_at").toLocalDateTime());
         return transaction;
     }
